fix (notification-service): change all ids type from int to long

+++ backend/notification-service/src/main/java/com/github/spjavaind300/notificationservice/service/CommentService.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 public interface CommentService {
 
-    List<Integer> getCommentsAuthorIdsOfAdv(int advId);
+    List<Long> getCommentsAuthorIdsOfAdv(long advId);
 
 }
